+++ codegen/src/main/kotlin/io/bkbn/skribe/codegen/domain/schema/SerializableSchema.kt
@@ -1,0 +1,8 @@
+package io.bkbn.skribe.codegen.domain.schema
+
+import com.squareup.kotlinpoet.TypeName
+
+sealed interface SerializableSchema {
+  val utilPackage: String
+  val serializerTypeName: TypeName
+}
