+++ codegen/src/main/kotlin/io/bkbn/skribe/codegen/domain/schema/SkribeEmailSchema.kt
@@ -1,0 +1,13 @@
+package io.bkbn.skribe.codegen.domain.schema
+
+import com.squareup.kotlinpoet.TypeName
+import com.squareup.kotlinpoet.asClassName
+import io.bkbn.skribe.codegen.domain.SkribeSpec
+
+data class SkribeEmailSchema(
+  override val name: String,
+  override val requiresSerialization: Boolean = false,
+) : SkribeScalarSchema {
+  context(SkribeSpec)
+  override fun toKotlinTypeName(): TypeName = String::class.asClassName()
+}
