+++ codegen/src/main/kotlin/io/bkbn/skribe/codegen/domain/schema/SkribePotentiallyScalarSchema.kt
@@ -1,0 +1,5 @@
+package io.bkbn.skribe.codegen.domain.schema
+
+sealed interface SkribePotentiallyScalarSchema : SkribeSchema {
+  fun isScalar(): Boolean
+}
