+++ codegen/src/main/kotlin/io/bkbn/skribe/codegen/domain/schema/SkribeScalarSchema.kt
@@ -1,0 +1,3 @@
+package io.bkbn.skribe.codegen.domain.schema
+
+sealed interface SkribeScalarSchema : SkribeSchema
